+++ pkg/agent/config.go
@@ -36,9 +36,6 @@ type Config struct {
 	// The TLS Certificate resource name to use for the default X509-SVID with Envoy SDS
 	DefaultSVIDName string
 
-	// Enable deprecated Envoy SDS v2 API
-	EnableDeprecatedSDSv2API bool
-
 	// If true, the agent will bootstrap insecurely with the server
 	InsecureBootstrap bool