+++ pkg/server/endpoints/bundle/config.go
@@ -1,6 +1,9 @@
 package bundle
 
-import "net"
+import (
+	"net"
+	"time"
+)
 
 type EndpointConfig struct {
 	// Address is the address on which to serve the federation bundle endpoint.
@@ -9,4 +12,6 @@ type EndpointConfig struct {
 	// ACME is the ACME configuration for the bundle endpoint.
 	// If unset, the bundle endpoint will use SPIFFE auth.
 	ACME *ACMEConfig
+
+	RefreshHint *time.Duration
 }