+++ pkg/server/server.go
@@ -208,7 +208,6 @@ func (s *Server) run(ctx context.Context) (err error) {
 		registrationManager.Run,
 		bundlePublishingManager.Run,
 		util.SerialRun(s.waitForTestDial, healthChecker.ListenAndServe),
-		scanForBadEntries(s.config.Log, metrics, cat.GetDataStore()),
 	}
 
 	if s.config.LogReopener != nil {
@@ -399,6 +398,7 @@ func (s *Server) newEndpointsServer(ctx context.Context, catalog catalog.Catalog
 	}
 	if s.config.Federation.BundleEndpoint != nil {
 		config.BundleEndpoint.Address = s.config.Federation.BundleEndpoint.Address
+		config.BundleEndpoint.RefreshHint = s.config.Federation.BundleEndpoint.RefreshHint
 		config.BundleEndpoint.ACME = s.config.Federation.BundleEndpoint.ACME
 	}
 	return endpoints.New(ctx, config)